Add files via upload
Rewrote asm to run under 32 bit, so that it worked on an old Debian Jessie. I'm not sure even running docker on it is ok, though so far so good I guess. Also no idea how git works really - I'm trying to push this to a forked version in my own repository and then get back to something more imnmediately useful.